backend/requirements.txt
# ComplianceX Backend Dependencies

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# AI / LLM
anthropic>=0.25.0

# Agent orchestration
langgraph>=0.1.0

# Vector store
chromadb>=0.5.0

# Environment variable loading
python-dotenv>=1.0.0

# HTTP client (used internally by anthropic SDK)
httpx>=0.27.0
